refactor: favorite 관련로직 FavoriteService로 이동
